+++ src/main/java/hk/org/ha/kcc/its/repository/EformRepository.java
@@ -3,5 +3,5 @@ package hk.org.ha.kcc.its.repository;
 import hk.org.ha.kcc.its.model.Eform;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EformRepository extends JpaRepository<Eform, Integer> {
+public interface EformRepository extends JpaRepository<Eform, Long> {
 }
